+++ js/escenas/menu-principal.js
@@ -321,7 +321,7 @@ export class MenuPrincipal {
     ctx.font = '11px monospace';
     ctx.fillStyle = '#333333';
     ctx.fillText(
-      'v0.8 - Liceo Frances de Santo Domingo 2026',
+      'v0.10 - Liceo Frances de Santo Domingo 2026',
       ancho / 2,
       alto - 15
     );